[BB-1244] Refactor after the review
